# legacy_openapi.yaml
openapi: 3.0.3
info:
  title: Portal API
  version: 1.0.0
  description: |
    REST API for Portal system.

    Scope:
      - Authentication
      - User profile
      - Admin user management
servers:
  - url: http://localhost:8000/api/v1
    description: Local
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful, verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid input
        '409':
          description: Email or username already exists / email is blacklisted
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email by token
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '201':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid or expired token
        '409':
          description: User not found or already deleted
        '500':
          description: Internal server error
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend verification email
      operationId: resendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
      responses:
        '201':
          description: Verification email resent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid input or token state
        '409':
          description: User not found or already deleted
        '500':
          description: Internal server error
  /auth/set-password:
    post:
      tags: [Auth]
      summary: Set an initial password using a password-set token
      operationId: setPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPasswordRequest'
      responses:
        '201':
          description: Password set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid token or password
        '409':
          description: User not found or already deleted
        '500':
          description: Internal server error
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using a password reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid token or password
        '409':
          description: User not found or already deleted
        '500':
          description: Internal server error
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Send a password reset email if the account exists
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Forgot password request processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '400':
          description: Invalid input
        '401':
          description: Invalid refresh token
        '500':
          description: Internal server error
...
  /auth/logout-all:
    post:
      tags: [Auth]
      summary: Logout all sessions for current user
      operationId: logoutAll
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout all successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /users/me:
    get:
      tags: [User]
      summary: Get current user profile
      operationId: getMyProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Internal server error
    put:
      tags: [User]
      summary: Update current user profile
      operationId: updateMyProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMyProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '409':
          description: Username already exists
        '500':
          description: Internal server error
  /users/me/change-password:
    put:
      tags: [User]
      summary: Change current user password
      operationId: changeMyPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMessageResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '409':
          description: Current password incorrect
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    RoleCode:
      type: string
      enum: [user, admin]
    UserStatus:
      type: string
      enum: [pending_verification, active, deleted]
    RegisterRequest:
      type: object
      required:
        - email
        - username
        - first_name
        - last_name
        - password
        - dob
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
          maxLength: 50
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 255
        dob:
          type: string
          format: date
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    SetPasswordRequest:
      type: object
      required:
        - token
        - password
        - confirm_password
      properties:
        token:
          type: string
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 255
        confirm_password:
          type: string
          format: password
          minLength: 8
          maxLength: 255
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
...
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    RefreshResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Access token expiry in seconds
...
    AuthMessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    UpdateMyProfileRequest:
      type: object
      properties:
        first_name:
          type: string
          maxLength: 100
          nullable: true
        last_name:
          type: string
          maxLength: 100
          nullable: true
        username:
          type: string
          minLength: 3
          maxLength: 50
          nullable: true
        dob:
          type: string
          format: date
          nullable: true
    ChangePasswordRequest:
      type: object
      required:
        - current_password
        - new_password
        - confirm_new_password
      properties:
        current_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
          minLength: 8
          maxLength: 255
        confirm_new_password:
          type: string
          format: password
          minLength: 8
          maxLength: 255
...
    UserResponse:
      type: object
      required:
        - id
        - email
        - username
        - first_name
        - last_name
        - role
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        dob:
          type: string
          format: date
          nullable: true
        role:
          $ref: '#/components/schemas/RoleCode'
        status:
          $ref: '#/components/schemas/UserStatus'
        email_verified_at:
          type: string
          format: date-time
          nullable: true
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
        deleted_by:
          type: string
          nullable: true
        restored_at:
          type: string
          format: date-time
          nullable: true
        restored_by:
          type: string
          format: uuid
          nullable: true
...